+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file. This change
 - Optionally add a `:resolver` key to state when validating a statement to inline resolver-fns prior to execution
 - Streamline statement validation output - remove :state key in the absence of errors
 - Add a spec for statement validation output and ensure test cases conform
+- Add starwars test case from the `graphql-clj-starter project`
 
 ## [0.1.19] - 2016-11-16
 - Support for Input Objects and their fields in Introspection Schema. By Boris Jonica (bjonica)